+++ lib/utils/network_utils.dart
@@ -25,6 +25,28 @@ import 'package:http/http.dart';
 import 'package:http/io_client.dart';
 import 'package:package_info/package_info.dart';
 
+/// Dummy network request can be used to trigger the network access permission
+/// on iOS devices. Doing this at an appropriate place in the code can prevent
+/// SocketExceptions.
+Future<void> dummyRequest({required Uri url, bool sslVerify = true}) async {
+  HttpClient httpClient = HttpClient();
+  httpClient.badCertificateCallback =
+      ((X509Certificate cert, String host, int port) => !sslVerify);
+  httpClient.userAgent = "privacyIDEA-App /"
+      " ${Platform.operatingSystem}"
+      " ${(await PackageInfo.fromPlatform()).version}";
+
+  IOClient ioClient = IOClient(httpClient);
+
+  try {
+    await ioClient.post(url, body: "");
+  } on SocketException {
+    // ignore
+  } finally {
+    ioClient.close();
+  }
+}
+
 /// Custom POST request allows to not verify certificates
 Future<Response> doPost(
     {required Uri url,
@@ -33,6 +55,18 @@ Future<Response> doPost(
   log("Sending post request",
       name: "utils.dart",
       error: "URI: $url, SSLVerify: $sslVerify, Body: $body");
+
+  List<MapEntry> entries =
+      body.entries.where((element) => element.value == null).toList();
+  if (entries.isNotEmpty) {
+    List<String> nullEntries = [];
+    for (MapEntry entry in entries) {
+      nullEntries.add(entry.key);
+    }
+    throw ArgumentError(
+        "Can not send request because the [body] contains a null values at entries $nullEntries,"
+        " this is not permitted.");
+  }
 
   HttpClient httpClient = HttpClient();
   httpClient.badCertificateCallback =