+++ lib/utils/parsing_utils.dart
@@ -271,28 +271,24 @@ Map<String, dynamic> parsePiAuth(Uri uri) {
         "[$pushVersionAsString] is not a valid value for parameter [v].");
   }
 
-  uriMap[URI_LABEL] = _parseLabel(uri);
+  List labelIssuerList = _parseLabelAndIssuer(uri);
+  uriMap[URI_LABEL] = labelIssuerList[0];
+  uriMap[URI_ISSUER] ??= labelIssuerList[1];
 
   uriMap[URI_SERIAL] = uri.queryParameters["serial"];
   ArgumentError.checkNotNull(uriMap[URI_SERIAL], "serial");
 
   uriMap[URI_PROJECT_ID] = uri.queryParameters["projectid"];
-  ArgumentError.checkNotNull(uriMap[URI_PROJECT_ID], "projectid");
 
   uriMap[URI_APP_ID] = uri.queryParameters["appid"];
-  ArgumentError.checkNotNull(uriMap[URI_APP_ID], "appid");
 
   uriMap[URI_APP_ID_IOS] = uri.queryParameters["appidios"];
-  ArgumentError.checkNotNull(uriMap[URI_APP_ID_IOS], "appidios");
 
   uriMap[URI_API_KEY] = uri.queryParameters["apikey"];
-  ArgumentError.checkNotNull(uriMap[URI_API_KEY], "apikey");
 
   uriMap[URI_API_KEY_IOS] = uri.queryParameters["apikeyios"];
-  ArgumentError.checkNotNull(uriMap[URI_API_KEY_IOS], "apikeyios");
 
   uriMap[URI_PROJECT_NUMBER] = uri.queryParameters["projectnumber"];
-  ArgumentError.checkNotNull(uriMap[URI_PROJECT_NUMBER], "projectnumber");
 
   String? url = uri.queryParameters["url"];
   ArgumentError.checkNotNull(url);
@@ -338,7 +334,9 @@ Map<String, dynamic> parseOtpAuth(Uri uri) {
     log("  $key | $value");
   });
 
-  uriMap[URI_LABEL] = _parseLabel(uri);
+  List labelIssuerList = _parseLabelAndIssuer(uri);
+  uriMap[URI_LABEL] = labelIssuerList[0];
+  uriMap[URI_ISSUER] ??= labelIssuerList[1];
 
   String algorithm = uri.queryParameters["algorithm"] ??
       enumAsString(Algorithms.SHA1); // Optional parameter
@@ -419,15 +417,10 @@ Map<String, dynamic> parseOtpAuth(Uri uri) {
   if (uriMap[URI_TYPE] == "totp") {
     // Parse period.
     String periodAsString = uri.queryParameters["period"] ?? "30";
-    if (periodAsString != "30" && periodAsString != "60") {
-      throw ArgumentError.value(
-        uri,
-        "uri",
-        "[$periodAsString] is not a valid value for the paramerter [period].",
-      );
-    }
 
-    uriMap[URI_PERIOD] = int.parse(periodAsString);
+    uriMap[URI_PERIOD] = int.parse(periodAsString,
+        onError: (e) => throw ArgumentError(
+            "Value [$periodAsString] for parameter [period] is invalid."));
   }
 
   if (is2StepURI(uri)) {
@@ -472,17 +465,26 @@ Map<String, dynamic> parseOtpAuth(Uri uri) {
   return uriMap;
 }
 
-String _parseLabel(Uri uri) {
-  String label;
+/// Parse the label and the issuer (if it exists) from the url.
+List _parseLabelAndIssuer(Uri uri) {
+  String label = "";
+  String issuer = "";
   String param = uri.path.substring(1);
+  param = Uri.decodeFull(param);
 
   try {
-    label = Uri.decodeFull(param);
+    if (param.contains(":")) {
+      List split = param.split(":");
+      issuer = split[0];
+      label = split[1];
+    } else {
+      label = param;
+    }
   } on Error {
     label = param;
   }
 
-  return label;
+  return [label, issuer];
 }
 
 String _parseIssuer(Uri uri) {