+++ lib/utils/storage_utils.dart
@@ -78,20 +78,20 @@ class StorageUtil {
 
       // TODO when the token version (token.version) changed handle this here.
 
-
       // Handle new fields here
       serializedToken['issuer'] ??= "";
       serializedToken['label'] ??= "";
 
-      if (serializedToken['type'] == enumAsString(TokenTypes.HOTP)) {
+      String type = serializedToken['type'];
+      if (type == enumAsString(TokenTypes.HOTP)) {
         tokenList.add(HOTPToken.fromJson(serializedToken));
-      } else if (serializedToken['type'] == enumAsString(TokenTypes.TOTP)) {
+      } else if (type == enumAsString(TokenTypes.TOTP)) {
         tokenList.add(TOTPToken.fromJson(serializedToken));
-      } else if (serializedToken['type'] == enumAsString(TokenTypes.PIPUSH)) {
+      } else if (type == enumAsString(TokenTypes.PIPUSH)) {
         tokenList.add(PushToken.fromJson(serializedToken));
       } else {
         log(
-          'Type ${serializedToken['type']} is unknown.',
+          'Type $type is unknown.',
           name: 'storage_utils.dart',
         );
       }
@@ -114,7 +114,8 @@ class StorageUtil {
   static const _GLOBAL_FIREBASE_CONFIG_KEY =
       _GLOBAL_PREFIX + "cc0d13b2-9ce1-11ea-bb37-0242ac130002";
 
-  static void saveOrReplaceGlobalFirebaseConfig(FirebaseConfig config) async =>
+  static Future<void> saveOrReplaceGlobalFirebaseConfig(
+          FirebaseConfig config) async =>
       await _storage.write(
           key: _GLOBAL_FIREBASE_CONFIG_KEY, value: jsonEncode(config));
 
@@ -192,7 +193,8 @@ class StorageUtil {
       Token token;
       String id = Uuid().v4();
 
-      if (tokenMap['type'] == 'hotp') {
+      String type = tokenMap['type'];
+      if (type == 'hotp') {
         token = HOTPToken(
           issuer: tokenMap['label'],
           id: id,
@@ -202,7 +204,7 @@ class StorageUtil {
           secret: tokenMap['secret'],
           algorithm: mapStringToAlgorithm(tokenMap['algorithm']),
         );
-      } else if (tokenMap['type'] == 'totp') {
+      } else if (type == 'totp') {
         token = TOTPToken(
           issuer: tokenMap['label'],
           id: id,
@@ -212,7 +214,7 @@ class StorageUtil {
           secret: tokenMap['secret'],
           algorithm: mapStringToAlgorithm(tokenMap['algorithm']),
         );
-      } else if (tokenMap['type'] == 'pipush') {
+      } else if (type == 'pipush') {
         token = PushToken(
           issuer: tokenMap['label'],
           label: tokenMap['label'],
@@ -242,8 +244,8 @@ class StorageUtil {
           projectNumber: configMap['projectnumber'],
         );
 
-        StorageUtil.saveOrReplaceFirebaseConfig(token, config);
-        StorageUtil.saveOrReplaceGlobalFirebaseConfig(config);
+        await StorageUtil.saveOrReplaceFirebaseConfig(token, config);
+        await StorageUtil.saveOrReplaceGlobalFirebaseConfig(config);
       } else {
         log(
           "Unknown token type encountered",
@@ -272,4 +274,10 @@ class StorageUtil {
   static Future<void> setCurrentVersion(String version) async {
     await _storage.write(key: _KEY_VERSION, value: version);
   }
+
+  // #########################################################################
+  // Misc
+  // #########################################################################
+
+  static Future<void> deleteEverything() async => _storage.deleteAll();
 }