+++ lib/widgets/token_widgets.dart
@@ -40,6 +40,7 @@ import 'package:privacyidea_authenticator/model/firebase_config.dart';
 import 'package:privacyidea_authenticator/model/tokens.dart';
 import 'package:privacyidea_authenticator/screens/main_screen.dart';
 import 'package:privacyidea_authenticator/utils/crypto_utils.dart';
+import 'package:privacyidea_authenticator/utils/identifiers.dart';
 import 'package:privacyidea_authenticator/utils/network_utils.dart';
 import 'package:privacyidea_authenticator/utils/parsing_utils.dart';
 import 'package:privacyidea_authenticator/utils/storage_utils.dart';
@@ -80,6 +81,26 @@ abstract class _TokenWidgetState extends State<TokenWidget> {
 
   _TokenWidgetState(this._token) {
     _saveThisToken();
+  }
+
+  List<Widget> _getSubtitle() {
+    List<Widget> children = [];
+    if (_token.label.isNotEmpty) {
+      children.add(Text(
+        _token.label,
+        style: Theme.of(context).textTheme.headline5,
+      ));
+    }
+    if (_token.issuer.isNotEmpty) {
+      children.add(Text(
+        _token.issuer,
+        style: Theme.of(context)
+            .textTheme
+            .headline6!
+            .copyWith(fontWeight: FontWeight.normal),
+      ));
+    }
+    return children;
   }
 
   @override
@@ -244,7 +265,7 @@ class _PushWidgetState extends _TokenWidgetState with LifecycleMixin {
 
     // Delete expired push requests periodically.
     _deleteTimer = Timer.periodic(Duration(seconds: 30), (_) {
-      if (_token.pushRequests != null && _token.pushRequests.isNotEmpty) {
+      if (_token.pushRequests.isNotEmpty) {
         _deleteExpiredRequests(_token);
         _saveThisToken();
 
@@ -319,9 +340,16 @@ class _PushWidgetState extends _TokenWidgetState with LifecycleMixin {
       setState(() => _rollOutFailed = false);
     }
 
+    // Trigger network permission request on iOS
+    if (Platform.isIOS) {
+      await dummyRequest(url: _token.url!, sslVerify: _token.sslVerify!);
+    }
+
+    FirebaseConfig tokenConfig =
+        (await StorageUtil.loadFirebaseConfig(_token))!;
     if (await StorageUtil.globalFirebaseConfigExists() &&
-        await StorageUtil.loadFirebaseConfig(_token) !=
-            await StorageUtil.loadGlobalFirebaseConfig()) {
+        tokenConfig.projectID != null && // Does not exist for poll only tokens
+        tokenConfig != await StorageUtil.loadGlobalFirebaseConfig()) {
       // The firebase config of this token is different to the existing
       // firebase config in this app.
       log("Token has different firebase config than existing.",
@@ -336,22 +364,6 @@ class _PushWidgetState extends _TokenWidgetState with LifecycleMixin {
         setState(() => _rollOutFailed = true);
       }
 
-      return;
-    }
-
-    if (DateTime.now().isAfter(_token.expirationDate)) {
-      log("Token is expired, abort roll-out and delete it.",
-          name: "token_widgets.dart",
-          error: "Now: ${DateTime.now()}, Token expires at ${[
-            _token.expirationDate
-          ]}, Token: $_token");
-
-      if (mounted) {
-        setState(() => _rollOutFailed = true);
-      }
-
-      _showMessage(
-          AppLocalizations.of(context)!.errorTokenExpired(_token.label), 3);
       return;
     }
 
@@ -370,6 +382,7 @@ class _PushWidgetState extends _TokenWidgetState with LifecycleMixin {
     }
 
     try {
+      // TODO What to do with poll only tokens if google-services is used?
       Response response =
           await doPost(sslVerify: _token.sslVerify!, url: _token.url!, body: {
         'enrollment_credential': _token.enrollmentCredentials,
@@ -404,6 +417,20 @@ class _PushWidgetState extends _TokenWidgetState with LifecycleMixin {
             AppLocalizations.of(context)!
                 .errorRollOutFailed(_token.label, response.statusCode),
             3);
+      }
+    } on PlatformException catch (e, s) {
+      log("Roll out push token [$_token] failed.",
+          name: "token_widgets.dart", error: e);
+
+      if (mounted) {
+        setState(() => _rollOutFailed = true);
+      }
+
+      if (e.code == FIREBASE_TOKEN_ERROR_CODE) {
+        _showMessage(
+            AppLocalizations.of(context)!.errorRollOutNoNetworkConnection, 3);
+      } else {
+        Catcher.reportCheckedError(e, s);
       }
     } on SocketException catch (e) {
       log("Roll out push token [$_token] failed.",
@@ -576,9 +603,10 @@ class _PushWidgetState extends _TokenWidgetState with LifecycleMixin {
                   _token.serial,
                   style: Theme.of(context).textTheme.headline4,
                 ),
-                subtitle: Text(
-                  _token.label,
-                  style: Theme.of(context).textTheme.headline5,
+                subtitle: Column(
+                  mainAxisAlignment: MainAxisAlignment.start,
+                  crossAxisAlignment: CrossAxisAlignment.start,
+                  children: _getSubtitle(),
                 ),
                 trailing: Icon(Icons.message),
               ),
@@ -767,9 +795,10 @@ class _HotpWidgetState extends _OTPTokenWidgetState {
             insertCharAt(_otpValue, " ", _token.digits ~/ 2),
             style: Theme.of(context).textTheme.headline4,
           ),
-          subtitle: Text(
-            _token.label,
-            style: Theme.of(context).textTheme.headline5,
+          subtitle: Column(
+            mainAxisAlignment: MainAxisAlignment.start,
+            crossAxisAlignment: CrossAxisAlignment.start,
+            children: _getSubtitle(),
           ),
         ),
         Align(
@@ -793,7 +822,7 @@ class _HotpWidgetState extends _OTPTokenWidgetState {
 class _TotpWidgetState extends _OTPTokenWidgetState
     with SingleTickerProviderStateMixin, LifecycleMixin {
   late AnimationController
-      controller; // Controller for animating the LinearProgressAnimator
+      _controller; // Controller for animating the LinearProgressAnimator
 
   TOTPToken get _token => super._token as TOTPToken;
 
@@ -818,7 +847,7 @@ class _TotpWidgetState extends _OTPTokenWidgetState
   void initState() {
     super.initState();
 
-    controller = AnimationController(
+    _controller = AnimationController(
       duration: Duration(seconds: _token.period),
       // Animate the progress for the duration of the tokens period.
       vsync: this,
@@ -831,7 +860,7 @@ class _TotpWidgetState extends _OTPTokenWidgetState
       ..addStatusListener((status) {
         // Add listener to restart the animation after the period, also updates the otp value.
         if (status == AnimationStatus.completed) {
-          controller.forward(from: _getCurrentProgress());
+          _controller.forward(from: _getCurrentProgress());
           _updateOtpValue();
         }
       })
@@ -844,12 +873,12 @@ class _TotpWidgetState extends _OTPTokenWidgetState
   @override
   void onResume() {
     _updateOtpValue();
-    controller.forward(from: _getCurrentProgress());
+    _controller.forward(from: _getCurrentProgress());
   }
 
   @override
   void dispose() {
-    controller.dispose(); // Dispose the controller to prevent memory leak.
+    _controller.dispose(); // Dispose the controller to prevent memory leak.
     super.dispose();
   }
 
@@ -862,13 +891,14 @@ class _TotpWidgetState extends _OTPTokenWidgetState
             insertCharAt(_otpValue, " ", _token.digits ~/ 2),
             style: Theme.of(context).textTheme.headline4,
           ),
-          subtitle: Text(
-            _token.label,
-            style: Theme.of(context).textTheme.headline5,
+          subtitle: Column(
+            mainAxisAlignment: MainAxisAlignment.start,
+            crossAxisAlignment: CrossAxisAlignment.start,
+            children: _getSubtitle(),
           ),
         ),
         LinearProgressIndicator(
-          value: controller.value,
+          value: _controller.value,
         ),
       ],
     );